переписал мвп под реакт

+++ Common/Requests/SimpleResearchRequest.cs
@@ -6,7 +6,7 @@ namespace PSSN.Common.Requests;
     Strats - выбранные стратегии
     A - матрица выйгришей
 */
-public record SimpleResearchRequest(int K,int R, IEnumerable<string> Strats,int[][] A)
+public record SimpleResearchRequest(int K,int R, IEnumerable<string> Strats,double[][] A)
 {
     
 }
